Performance improvements

+++ src/sd_driver/SDIO/SdioCard.h
@@ -145,7 +145,7 @@ uint32_t sd_sdio_status(sd_card_t *sd_card_p);
  * \return The number of 512 byte data sectors in the card
  *         or zero if an error occurs.
  */
-uint64_t sd_sdio_sectorCount(sd_card_t *sd_card_p);
+uint32_t sd_sdio_sectorCount(sd_card_t *sd_card_p);
 /**
  *  Send CMD12 to stop read or write.
  *
